Breaks Footer into separate components
