+++ src/App.jsx
@@ -6,12 +6,12 @@ import About from './About.jsx';
 import Header from './Header.jsx';
 import Contact from './Contact.jsx';
 import TopNav from './TopNav.jsx';
-import Footer from './Footer.jsx';
+import Footer from './Footer';
 import { ThemeProvider, createMuiTheme } from '@material-ui/core/styles';
-
 
 const App = () => {
     const classes = useStyles();
+
     const theme = createMuiTheme({
         typography: {
             fontFamily: ['Poppins, sans-serif'],
@@ -21,52 +21,35 @@ const App = () => {
         }
     });
 
-
     return (
         <>
         <ThemeProvider theme={theme}>
             <CssBaseline />
         {/*------------------------------Navigation Section------------------------------*/}
-
-         
             <TopNav />
-
         {/*------------------------------Header Section------------------------------*/}
-
             <div className={classes.main} id='main'>
                 <Header />
             </div>
-
         {/*------------------------------About Section------------------------------*/}
-
             <div className={classes.about} id='about'>
                 <About />
             </div>
-
         {/*------------------------------Testimonials Section------------------------------*/}
-
             <div className={classes.testimonials} id='testimonials'>
                 <Testimonials />
             </div>
-
         {/*------------------------------Contact Section------------------------------*/}
-
             <div className={classes.contact} id='contact'>
                 <Contact />
             </div>
-
         {/*------------------------------Footer Section------------------------------*/}
             <div>
                 <Footer />
             </div>
-        
-        
-        
         </ThemeProvider>
         </>
     )
 }
-
-
 
 export default App;
