+++ src/Footer/logoArray.js
@@ -1,0 +1,28 @@
+import DBL from '../assets/logos/DoverBigLocal.png';
+import GBCA from '../assets/logos/GBCAJudge.jpg';
+import NCVO from '../assets/logos/NCVOLogo.png';
+import NAG from '../assets/logos/nestAndGrow.png';
+import TSL from '../assets/logos/ThirdSectorLeaders.png';
+
+export default [
+    {
+        logoName: 'Dover Big Local',
+        path: DBL,
+    },
+    {
+        logoName: 'Great British Care Awards',
+        path: GBCA,
+    },
+    {
+        logoName: 'National Council for Voluntary Organisations',
+        path: NCVO,
+    },
+    {
+        logoName: 'Nest And Grow',
+        path: NAG,
+    },
+    {
+        logoName: 'Third Sector Leaders',
+        path: TSL,
+    }
+];
