--- src/Theme.js
@@ -1,9 +1,0 @@
-import { createMuiTheme } from '@material-ui/core/styles';
-
-const theme = createMuiTheme({
-    typography: {
-        fontFamily: 'Roboto',
-    },
-})
-
-export default theme;
