--- src/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { ThemeProvider } from '@material-ui/core/styles';
-import Theme from './Theme';
-
-import App from './App';
-
-ReactDOM.render(
-    <ThemeProvider theme={Theme}>
-        <App />
-    </ThemeProvider>, 
-document.getElementById('root'));
