+++ src/styles.js
@@ -1,5 +1,5 @@
 import { makeStyles } from '@material-ui/core/styles';
-import logo from './MMLogo.png';
+import logo from './assets/MMLogo.png';
 
 const useStyles = makeStyles((theme) => ({
     appBar: {
